// meridian-ui/src/lib/button/button.enums.ts
export enum ButtonVariant {
  Filled = 'filled',
  Outline = 'outline',
  Ghost = 'ghost',
  Link = 'link',
}

export enum ButtonColor {
  Primary = 'primary',
  Secondary = 'secondary',
  Neutral = 'neutral',
  Success = 'success',
  Warning = 'warning',
  Error = 'error',
  Info = 'info',
}

export enum ButtonSize {
  Xs = 'xs',
  Sm = 'sm',
  Md = 'md',
  Lg = 'lg',
  Xl = 'xl',
}

export enum ButtonShape {
  Default = 'default',
  Square = 'square',
}

export enum ButtonRadius {
  None = 'none',
  Sm = 'sm',
  Md = 'md',
  Lg = 'lg',
  Pill = 'pill',
}

export enum ButtonStatus {
  Default = 'default',
  Loading = 'loading',
}